+++ src/stories/Footer.stories.ts
@@ -19,6 +19,5 @@ export const Primary: Story = {
   args: {
     path: '/',
     fixed: true,
-    desktop: false,
   },
 };
