+++ src/utils/useIsDesktop.ts
@@ -1,0 +1,24 @@
+'use client';
+
+import { useState, useEffect } from 'react';
+
+export default function useIsDesktop() {
+  const [isDesktop, setIsDesktop] = useState(true);
+
+  useEffect(() => {
+    const handleResize = () => {
+      const width = window.innerWidth;
+      setIsDesktop(width > 768);
+    };
+
+    handleResize(); // Call the function initially
+
+    window.addEventListener('resize', handleResize); // Add event listener for window resize
+
+    return () => {
+      window.removeEventListener('resize', handleResize); // Clean up the event listener on component unmount
+    };
+  }, []);
+
+  return isDesktop;
+}
